Introduction To Hierarchical Clustering with R
This basic introduction to hierarchical clustering with R, should
highlight how "average linkage" may be inappropriate for grouping data.
It includes example code to generate the data, and examine the effects
of different linkage methods on the clades formed.

The scoring section shows methods for examining the validity of a
cluster which are less commonly found than the "elbow method". The
output of a measurable score for the allows for testing on improvements
of cluster quality.